requirements.txt
openai>=1.0.0
textual>=0.62.0
python-dotenv>=1.0.0
